pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signal-engine"
version = "0.1.0"
description = "Python collect CLI for signal-oriented collection lanes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
signal-engine = "signal_engine.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
